Move fetch_signed_resources to utils
