+++ checks/remotesettings/signatures_age.py
@@ -6,6 +6,8 @@ from datetime import datetime, timezone
 
 from kinto_http import Client, BearerTokenAuth
 
+from .utils import fetch_signed_resources
+
 
 def utcnow():
     # Tiny wrapper, used for mocking in tests.
@@ -24,36 +26,6 @@ def get_signature_age_hours(client, bucket, collection):
     return age
 
 
-def fetch_source_collections(client):
-    # List signed collection using capabilities.
-    info = client.server_info()
-    try:
-        resources = info["capabilities"]["signer"]["resources"]
-    except KeyError:
-        raise ValueError("No signer capabilities found. Run on *writer* server!")
-
-    collections = set()
-    monitored = client.get_records(
-        bucket="monitor", collection="changes", _sort="bucket,collection"
-    )
-    for entry in monitored:
-        bid = entry["bucket"]
-        cid = entry["collection"]
-
-        for resource in resources:
-            dest = resource["destination"]
-            if dest["bucket"] == bid:
-                if dest["collection"] == cid or dest["collection"] is None:
-                    collections.add(
-                        (
-                            resource["source"]["bucket"],
-                            resource["source"]["collection"] or cid,
-                        )
-                    )
-
-    return sorted(collections)
-
-
 # TODO: should retry requests. cf. lambdas code
 async def run(query, server, auth, max_age):
     max_age = int(query.get("max_age", max_age))
@@ -67,15 +39,20 @@ async def run(query, server, auth, max_age):
     )
     client = Client(server_url=server, auth=auth)
 
-    collections = fetch_source_collections(client)
+    source_collections = [
+        (r["source"]["bucket"], r["source"]["collection"])
+        for r in fetch_signed_resources(server, auth)
+    ]
 
     loop = asyncio.get_event_loop()
     futures = [
         loop.run_in_executor(None, get_signature_age_hours, client, bid, cid)
-        for (bid, cid) in collections
+        for (bid, cid) in source_collections
     ]
     results = await asyncio.gather(*futures)
 
-    ages = {f"{bid}/{cid}": age for ((bid, cid), age) in zip(collections, results)}
+    ages = {
+        f"{bid}/{cid}": age for ((bid, cid), age) in zip(source_collections, results)
+    }
     all_good = all([age is not None and age < max_age for age in ages.values()])
     return all_good, ages
