+++ checks/remotesettings/utils.py
@@ -1,0 +1,54 @@
+import copy
+
+from kinto_http import Client
+
+
+def fetch_signed_resources(server_url, auth):
+    # List signed collection using capabilities.
+    client = Client(server_url=server_url, auth=auth)
+    info = client.server_info()
+    try:
+        resources = info["capabilities"]["signer"]["resources"]
+    except KeyError:
+        raise ValueError("No signer capabilities found. Run on *writer* server!")
+
+    # Build the list of signed collections, source -> preview -> destination
+    # For most cases, configuration of signed resources is specified by bucket and
+    # does not contain any collection information.
+    resources_by_bid = {}
+    resources_by_cid = {}
+    preview_buckets = set()
+    for resource in resources:
+        bid = resource["destination"]["bucket"]
+        cid = resource["destination"]["collection"]
+        if resource["source"]["collection"] is not None:
+            resources_by_cid[(bid, cid)] = resource
+        else:
+            resources_by_bid[bid] = resource
+        if "preview" in resource:
+            preview_buckets.add(resource["preview"]["bucket"])
+
+    resources = []
+    monitored = client.get_records(
+        bucket="monitor", collection="changes", _sort="bucket,collection"
+    )
+    for entry in monitored:
+        bid = entry["bucket"]
+        cid = entry["collection"]
+
+        # Skip preview collections entries
+        if bid in preview_buckets:
+            continue
+
+        if (bid, cid) in resources_by_cid:
+            r = resources_by_cid[(bid, cid)]
+        elif bid in resources_by_bid:
+            r = copy.deepcopy(resources_by_bid[bid])
+            r["source"]["collection"] = r["destination"]["collection"] = cid
+            if "preview" in r:
+                r["preview"]["collection"] = cid
+        else:
+            raise ValueError(f"Unknown signed collection {bid}/{cid}")
+        resources.append(r)
+
+    return resources
